requirements-mcp.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
mcp>=1.0.0
pinecone[grpc]>=5.0.0
python-dotenv>=1.0.0
sentence-transformers>=3.0.0
pydantic>=2.0.0
sse-starlette>=2.0.0
